% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sf-workman.R
\name{geom_sf_workman}
\alias{geom_sf_workman}
\title{Workman wrapper for geom_sf}
\usage{
geom_sf_workman(
  mapping = NULL,
  data = NULL,
  stat = "sf",
  position = "identity",
  ...,
  colour = workman_neutrals("base_200"),
  linewidth = 0.15,
  alpha = NA,
  inherit.aes = TRUE,
  show.legend = NA
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes().}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use.}

\item{position}{Position adjustment.}

\item{...}{Additional arguments passed to ggplot2::geom_sf().}

\item{colour}{Border colour for polygons.}

\item{linewidth}{Border line width.}

\item{alpha}{Alpha transparency.}

\item{inherit.aes}{Whether to inherit aesthetics.}

\item{show.legend}{Whether to include the layer in legends.}
}
\description{
A convenience wrapper around ggplot2::geom_sf() that applies
default border styling for choropleth and administrative maps.
}
