% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-data.R
\name{get_workman_seq}
\alias{get_workman_seq}
\title{Internal helper to get sequential palette}
\usage{
get_workman_seq(palette = "blue")
}
\arguments{
\item{palette}{One of "blue", "green", or "red".}
}
\description{
Internal helper to get sequential palette
}
\keyword{internal}
