% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_workman_fonts}
\alias{register_workman_fonts}
\title{Register Google fonts used by workmanThemes}
\usage{
register_workman_fonts()
}
\description{
Downloads and registers the Google Fonts used in the package.
}
