% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_colour_workman_b}
\alias{scale_colour_workman_b}
\alias{scale_color_workman_b}
\title{Binned colour scale}
\usage{
scale_colour_workman_b(palette = "blue", ...)

scale_color_workman_b(palette = "blue", ...)
}
\arguments{
\item{palette}{One of "blue", "green", or "red".}

\item{...}{Passed to ggplot2::scale_colour_stepsn().}
}
\description{
Binned colour scale
}
