% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_colour_workman_c}
\alias{scale_colour_workman_c}
\alias{scale_color_workman_c}
\title{Continuous colour scale}
\usage{
scale_colour_workman_c(palette = "blue", ...)

scale_color_workman_c(palette = "blue", ...)
}
\arguments{
\item{palette}{One of "blue", "green", or "red".}

\item{...}{Passed to ggplot2::scale_colour_gradientn().}
}
\description{
Continuous colour scale
}
