% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-discrete.R
\name{scale_colour_workman_d}
\alias{scale_colour_workman_d}
\alias{scale_color_workman_d}
\title{Discrete colour scale}
\usage{
scale_colour_workman_d(...)

scale_color_workman_d(...)
}
\arguments{
\item{...}{Passed to ggplot2::discrete_scale().}
}
\description{
Discrete colour scale
}
