% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_colour_workman_div}
\alias{scale_colour_workman_div}
\alias{scale_color_workman_div}
\title{Diverging colour scale}
\usage{
scale_colour_workman_div(midpoint = 0, ...)

scale_color_workman_div(midpoint = 0, ...)
}
\arguments{
\item{midpoint}{Midpoint for diverging scale.}

\item{...}{Passed to ggplot2::scale_colour_gradient2().}
}
\description{
Diverging colour scale
}
