% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_fill_workman_b}
\alias{scale_fill_workman_b}
\title{Binned fill scale}
\usage{
scale_fill_workman_b(palette = "blue", ...)
}
\arguments{
\item{palette}{One of "blue", "green", or "red".}

\item{...}{Passed to ggplot2::scale_fill_stepsn().}
}
\description{
Binned fill scale
}
