% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-discrete.R
\name{scale_fill_workman_d}
\alias{scale_fill_workman_d}
\title{Discrete fill scale}
\usage{
scale_fill_workman_d(...)
}
\arguments{
\item{...}{Passed to ggplot2::discrete_scale().}
}
\description{
Discrete fill scale
}
