% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-continuous.R
\name{scale_fill_workman_div}
\alias{scale_fill_workman_div}
\title{Diverging fill scale}
\usage{
scale_fill_workman_div(midpoint = 0, ...)
}
\arguments{
\item{midpoint}{Midpoint for diverging scale.}

\item{...}{Passed to ggplot2::scale_fill_gradient2().}
}
\description{
Diverging fill scale
}
