% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-map.R
\name{scale_fill_workman_map_b}
\alias{scale_fill_workman_map_b}
\title{Binned fill scale for maps}
\usage{
scale_fill_workman_map_b(
  palette = "blue",
  n.breaks = 6,
  labels = scales::label_number(),
  ...
)
}
\arguments{
\item{palette}{One of "blue", "green", or "red".}

\item{n.breaks}{Approximate number of bins.}

\item{labels}{A labelling function, e.g. scales::label_number(),
scales::label_dollar(), or scales::label_percent(scale = 1).}

\item{...}{Passed to ggplot2::scale_fill_stepsn().}
}
\description{
Convenience fill scale for choropleth maps using workmanThemes defaults.
}
