% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-map.R
\name{scale_fill_workman_map_c}
\alias{scale_fill_workman_map_c}
\title{Continuous fill scale for maps}
\usage{
scale_fill_workman_map_c(
  palette = "blue",
  n.breaks = 6,
  labels = scales::label_number(),
  ...
)
}
\arguments{
\item{palette}{One of "blue", "green", or "red".}

\item{n.breaks}{Approximate number of legend breaks.}

\item{labels}{A labelling function, e.g. scales::label_number(),
scales::label_dollar(), or scales::label_percent(scale = 1).}

\item{...}{Passed to ggplot2::scale_fill_gradientn().}
}
\description{
Convenience continuous fill scale for choropleth maps using
workmanThemes defaults.
}
