% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-labels.R
\name{scale_x_dollar_workman}
\alias{scale_x_dollar_workman}
\title{Dollar labels for x axis}
\usage{
scale_x_dollar_workman(prefix = "$", suffix = "", scale = 1, accuracy = 1, ...)
}
\arguments{
\item{prefix}{Prefix for dollar labels.}

\item{suffix}{Suffix for labels.}

\item{scale}{Scale factor.}

\item{accuracy}{Label accuracy.}

\item{...}{Passed to ggplot2::scale_x_continuous().}
}
\description{
Dollar labels for x axis
}
