% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-labels.R
\name{scale_y_percent_workman}
\alias{scale_y_percent_workman}
\title{Percent labels for y axis}
\usage{
scale_y_percent_workman(scale = 100, accuracy = 1, ...)
}
\arguments{
\item{scale}{Scale factor for percent labels.}

\item{accuracy}{Label accuracy.}

\item{...}{Passed to ggplot2::scale_x_continuous().}
}
\description{
Percent labels for y axis
}
