% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_workman.R
\name{theme_workman}
\alias{theme_workman}
\title{Workman ggplot2 theme}
\usage{
theme_workman(
  base_size = 12,
  base_family = "Alegreya Sans",
  title_family = "Montserrat",
  subtitle_family = "Open Sans",
  caption_family = "Inconsolata",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base body font family.}

\item{title_family}{Font family for plot titles.}

\item{subtitle_family}{Font family for subtitles.}

\item{caption_family}{Font family for captions.}

\item{base_line_size}{Base line size.}

\item{base_rect_size}{Base rectangle line size.}
}
\description{
Clean publication theme with separate fonts for title, subtitle,
body, and caption text.
}
