% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_workman.R
\name{theme_workman_map}
\alias{theme_workman_map}
\title{Workman map theme}
\usage{
theme_workman_map(
  base_size = 12,
  base_family = "Alegreya Sans",
  title_family = "Montserrat",
  subtitle_family = "Open Sans",
  caption_family = "Inconsolata"
)
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base body font family.}

\item{title_family}{Font family for plot titles.}

\item{subtitle_family}{Font family for subtitles.}

\item{caption_family}{Font family for captions.}
}
\description{
Variant of the workman theme tuned for choropleths and spatial graphics.
}
