% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workmanThemes-package.R
\docType{package}
\name{workmanThemes-package}
\alias{workmanThemes}
\alias{workmanThemes-package}
\title{workmanThemes: Personal ggplot2 Themes and Scales}
\description{
Publication-oriented ggplot2 themes, Flexoki-inspired scales,
Google font helpers, map themes, and label utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamuelWorkman/workmanThemes}
  \item Report bugs at \url{https://github.com/SamuelWorkman/workmanThemes/issues}
}

}
\author{
\strong{Maintainer}: Samuel Workman \email{sam.workman@mail.wvu.edu}

}
\keyword{internal}
